+++ MSAL/src/native_auth/public/MSALNativeAuthPublicClientApplication.swift
@@ -149,26 +149,20 @@ public final class MSALNativeAuthPublicClientApplication: MSALPublicClientApplic
 
     // MARK: delegate methods
 
-    /// Sign up a user with a given username and password.
+    /// Sign up a user using parameters.
     /// - Parameters:
-    ///   - username: Username for the new account.
-    ///   - password: Optional. Password to be used for the new account.
-    ///   - attributes: Optional. User attributes to be used during account creation.
-    ///   - correlationId: Optional. UUID to correlate this request with the server for debugging.
+    ///   - parameters: Parameters used for the Sign Up flow.
     ///   - delegate: Delegate that receives callbacks for the Sign Up flow.
     public func signUp(
-        username: String,
-        password: String? = nil,
-        attributes: [String: Any]? = nil,
-        correlationId: UUID? = nil,
+        parameters: MSALNativeAuthSignUpParameters,
         delegate: SignUpStartDelegate
     ) {
         Task {
             let controllerResponse = await signUpInternal(
-                username: username,
-                password: password,
-                attributes: attributes,
-                correlationId: correlationId
+                username: parameters.username,
+                password: parameters.password,
+                attributes: parameters.attributes,
+                correlationId: parameters.correlationId
             )
 
             let delegateDispatcher = SignUpStartDelegateDispatcher(delegate: delegate, telemetryUpdate: controllerResponse.telemetryUpdate)
@@ -190,26 +184,49 @@ public final class MSALNativeAuthPublicClientApplication: MSALPublicClientApplic
         }
     }
 
-    /// Sign in a user with a given username and password.
+    /// Sign up a user with a given username and password.
     /// - Parameters:
-    ///   - username: Username for the account
-    ///   - password: Optional. Password for the account.
-    ///   - scopes: Optional. Permissions you want included in the access token received after sign in flow has completed.
+    ///   - username: Username for the new account.
+    ///   - password: Optional. Password to be used for the new account.
+    ///   - attributes: Optional. User attributes to be used during account creation.
     ///   - correlationId: Optional. UUID to correlate this request with the server for debugging.
-    ///   - delegate: Delegate that receives callbacks for the Sign In flow.
-    public func signIn(
+    ///   - delegate: Delegate that receives callbacks for the Sign Up flow.
+    @available(*, deprecated, message: "This method is now deprecated. Use the method 'signUp(parameters:)' instead.")
+    public func signUp(
         username: String,
         password: String? = nil,
-        scopes: [String]? = nil,
+        attributes: [String: Any]? = nil,
         correlationId: UUID? = nil,
+        delegate: SignUpStartDelegate
+    ) {
+        Task {
+            let parameters = MSALNativeAuthSignUpParameters(username: username)
+            parameters.password = password
+            parameters.attributes = attributes
+            parameters.correlationId = correlationId
+            signUp(
+                parameters: parameters,
+                delegate: delegate
+            )
+        }
+    }
+
+    /// Sign in a user using parameters.
+    /// - Parameters:
+    ///   - parameters: Parameters used for the Sign In flow.
+    ///   - delegate: Delegate that receives callbacks for the Sign In flow.
+    public func signIn(
+        parameters: MSALNativeAuthSignInParameters,
         delegate: SignInStartDelegate
     ) {
         Task {
+            let claimsRequestJson = parameters.claimsRequest?.jsonString()
             let controllerResponse = await signInInternal(
-                username: username,
-                password: password,
-                scopes: scopes,
-                correlationId: correlationId
+                username: parameters.username,
+                password: parameters.password,
+                scopes: parameters.scopes,
+                claimsRequestJson: claimsRequestJson,
+                correlationId: parameters.correlationId
             )
 
             let delegateDispatcher = SignInStartDelegateDispatcher(delegate: delegate, telemetryUpdate: controllerResponse.telemetryUpdate)
@@ -235,18 +252,42 @@ public final class MSALNativeAuthPublicClientApplication: MSALPublicClientApplic
         }
     }
 
-    /// Reset the password for a given username.
+    /// Sign in a user with a given username and password.
     /// - Parameters:
-    ///   - username: Username for the account.
+    ///   - username: Username for the account
+    ///   - password: Optional. Password for the account.
+    ///   - scopes: Optional. Permissions you want included in the access token received after sign in flow has completed.
     ///   - correlationId: Optional. UUID to correlate this request with the server for debugging.
+    ///   - delegate: Delegate that receives callbacks for the Sign In flow.
+    @available(*, deprecated, message: "This method is now deprecated. Use the method 'signIn(parameters:)' instead.")
+    public func signIn(
+        username: String,
+        password: String? = nil,
+        scopes: [String]? = nil,
+        correlationId: UUID? = nil,
+        delegate: SignInStartDelegate
+    ) {
+        let parameters = MSALNativeAuthSignInParameters(username: username)
+        parameters.password = password
+        parameters.scopes = scopes
+        parameters.correlationId = correlationId
+        signIn(
+            parameters: parameters,
+            delegate: delegate
+        )
+    }
+
+    /// Reset the password using parameters
+    /// - Parameters:
+    ///   - parameters: Parameters used for the Reset Password flow.
     ///   - delegate: Delegate that receives callbacks for the Reset Password flow.
     public func resetPassword(
-        username: String,
-        correlationId: UUID? = nil,
+        parameters: MSALNativeAuthResetPasswordParameters,
         delegate: ResetPasswordStartDelegate
     ) {
         Task {
-            let controllerResponse = await resetPasswordInternal(username: username, correlationId: correlationId)
+            let controllerResponse = await resetPasswordInternal(username: parameters.username,
+                                                                 correlationId: parameters.correlationId)
 
             let delegateDispatcher = ResetPasswordStartDelegateDispatcher(delegate: delegate, telemetryUpdate: controllerResponse.telemetryUpdate)
 
@@ -263,6 +304,25 @@ public final class MSALNativeAuthPublicClientApplication: MSALPublicClientApplic
                 await delegate.onResetPasswordStartError(error: error)
             }
         }
+    }
+
+    /// Reset the password for a given username.
+    /// - Parameters:
+    ///   - username: Username for the account.
+    ///   - correlationId: Optional. UUID to correlate this request with the server for debugging.
+    ///   - delegate: Delegate that receives callbacks for the Reset Password flow.
+    @available(*, deprecated, message: "This method is now deprecated. Use the method 'resetPassword(parameters:)' instead.")
+    public func resetPassword(
+        username: String,
+        correlationId: UUID? = nil,
+        delegate: ResetPasswordStartDelegate
+    ) {
+        let parameters = MSALNativeAuthResetPasswordParameters(username: username)
+        parameters.correlationId = correlationId
+        resetPassword(
+            parameters: parameters,
+            delegate: delegate
+        )
     }
 
     /// Retrieve the current signed in account from the cache.