+++ MSAL/test/unit/native_auth/public/delegate/mfa/MFAGetAuthMethodsDelegateDispatcherTests.swift
@@ -48,7 +48,7 @@ final class MFAGetAuthMethodsDelegateDispatcherTests: XCTestCase {
             self.telemetryExp.fulfill()
         })
 
-        let expectedState = MFARequiredState(controller: controllerFactoryMock.signInController, scopes: [], continuationToken: "continuationToken", correlationId: correlationId)
+        let expectedState = MFARequiredState(controller: controllerFactoryMock.signInController, scopes: [], claimsRequestJson: nil, continuationToken: "continuationToken", correlationId: correlationId)
         let expectedAuthMethods = [MSALAuthMethod(id: "1", challengeType: "oob", loginHint: "us**@**oso.com", channelTargetType: MSALNativeAuthChannelType(value: "email"))]
 
         await sut.dispatchSelectionRequired(authMethods: expectedAuthMethods, newState: expectedState, correlationId: correlationId)
@@ -76,7 +76,7 @@ final class MFAGetAuthMethodsDelegateDispatcherTests: XCTestCase {
             self.telemetryExp.fulfill()
         })
 
-        let expectedState = MFARequiredState(controller: controllerFactoryMock.signInController, scopes: [], continuationToken: "continuationToken", correlationId: correlationId)
+        let expectedState = MFARequiredState(controller: controllerFactoryMock.signInController, scopes: [], claimsRequestJson: nil, continuationToken: "continuationToken", correlationId: correlationId)
 
         await sut.dispatchSelectionRequired(authMethods: [], newState: expectedState, correlationId: correlationId)